Start breaking apart WorldManager into classic map data handler and general multiworld management
